+++ index.js
@@ -1,10 +1,10 @@
 #!/usr/bin/env node
 const fs = require('fs');
 const path = require('path');
-const {exec} = require('child_process');
 const Rsync = require('rsync');
-const {log, settings, deploy} = require("./helpers/ops");
-let project = settings(process.argv[2]||`${process.cwd()}/package.json`);
+const {settings, deploy} = require("./scripts/ops");
+
+let project = settings(path.resolve(process.cwd(), process.argv[2])||path.resolve(process.cwd(),"./package.json"));
 
 try {
 	let {
@@ -21,7 +21,7 @@ try {
 		let n = Math.abs(template_input.length - template_output.length);
 		throw `I/O mismatch:\n\t${n<0?"template_input":"template_output"} needs ${n} more element${n>0?"s":""}`;
 	}
-	output = output.undot();
+	// output = path.resolve(process.cwd(),output);
 	outlet = outlet.replace("{output}", output);
 	assets_dir =  assets.replace("{outlet}", outlet)
 	views_dir = views.replace("{outlet}", outlet);
