+++ package.json
@@ -1,18 +1,17 @@
 {
-  "name": "front-to-back",
-  "version": "1.0.0",
-  "description": "Simple script to deploy and template index.html to another filetype",
-  "main": "index.js",
-  "bin":{
-    "f2b": "./index.js"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "suli-g <suligzus@gmail.com>",
-  "license": "MIT",
-  "dependencies": {
-    "f2b": "^1.0.0",
-    "rsync": "^0.6.1"
-  }
+    "name": "front-to-back",
+    "version": "0.1.2",
+    "description": "Simple script to deploy and template index.html to another filetype",
+    "main": "index.js",
+    "bin": {
+        "f2b": "./index.js"
+    },
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "suli-g <suligzus@gmail.com>",
+    "license": "MIT",
+    "dependencies": {
+        "rsync": "^0.6.1"
+    }
 }
