--- project.json
@@ -1,10 +1,0 @@
-{
-    "src": ".",
-    "outlet":"../msauwc",
-    "build": "build",
-    "views": "{outlet}/application/views/",
-    "assets": "{outlet}/assets",
-    "index_file": "dashboard.php",
-    "template_input": [["__([a-zA-Z]\\w*)__"], [">|<","[\\w -]+"], ["|...", "[a-zA-Z]\\w*"] ],
-    "template_output": [ ["<?=$ ", "?>"],["<?php ","?>"], ["<?php foreach ($", "as $key=>$value):?>"]]
-}
