+++ scripts/init.js
