+++ scripts/log.js
