+++ scripts/reg_val.js
