+++ scripts/sync.js
